Reorganize affordances Indicator
Rename, states for button, move to affordances folder.
