+++ src/Components/Actions/ReaderSettings/PlaygroundReaderSettingsContainer.tsx
@@ -16,7 +16,7 @@ import { PlaygroundPaginatedAffordancesGroup } from "@/Components/Settings/Reade
 import { PlaygroundScrollAffordancesGroup } from "@/Components/Settings/Reader/Affordances/Scroll/PlaygroundScrollAffordancesGroup";
 import { PlaygroundPaginatedAffordancesGroupContainer } from "@/Components/Settings/Reader/Affordances/Paginated/PlaygroundPaginatedAffordancesGroup";
 import { PlaygroundScrollAffordancesGroupContainer } from "@/Components/Settings/Reader/Affordances/Scroll/PlaygroundScrollAffordancesGroup";
-import { PlaygroundAffordancesIndicator } from "@/Components/Settings/PlaygroundAffordancesIndicator";
+import { PlaygroundAffordancesIndicator } from "@/Components/Settings/Reader/Affordances/PlaygroundAffordancesIndicator";
 
 import { 
   StatefulActionContainerProps,
